requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
Django==4.2.20
django-cors-headers==4.7.0
django-mongoengine==0.5.6
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dnspython==2.7.0
dotenv==0.9.9
gunicorn==23.0.0
idna==3.10
MarkupSafe==3.0.2
mongoengine==0.29.1
msgpack==1.1.0
packaging==24.2
PyJWT==2.9.0
pymongo==4.11.3
PyMySQL==1.1.1
python-dotenv==1.0.1
pytz==2025.1
redis==5.2.1
requests==2.32.3
sqlparse==0.5.3
urllib3==2.3.0
Werkzeug==3.1.3
whitenoise==6.9.0
